+++ packages/core/src/go.ts
@@ -203,7 +203,7 @@ export type { UndoManagerEvent, UndoManagerEventHandler } from './undo/UndoManag
 // Undo
 export { UndoManager } from './undo/UndoManager.ts';
 
-export const version = '1.8.0';
+export const version = '1.9.0';
 
 export type { ContextMenuItem, ContextMenuOptions } from './export/ContextMenu.ts';
 export { ContextMenu, createDefaultContextMenu } from './export/ContextMenu.ts';