+++ packages/core/src/vue/index.ts
@@ -218,4 +218,4 @@ export const Overview = defineComponent({
   },
 });
 
-export const version: string = '1.8.0';
+export const version: string = '1.9.0';