+++ packages/core/test/react.test.tsx
@@ -56,7 +56,7 @@ function renderStateful<P>(
 
 describe('graphojs/react', () => {
   it('exposes a version', () => {
-    expect(version).toBe('1.8.0');
+    expect(version).toBe('1.9.0');
   });
 
   it('renders a Diagram component and initializes the diagram', async () => {